requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
gast==0.3.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.7.1
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.16.6
opencv-python==4.1.2.30
pbr==5.4.4
pkg-resources==0.0.0
protobuf==3.11.2
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3
scipy==1.4.1
six==1.13.0
stevedore==1.31.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
Werkzeug==0.16.0
wrapt==1.11.2
